+++ src/main/java/aurora/carevisionapiserver/global/auth/util/RefreshTokenValidator.java
@@ -1,10 +1,9 @@
-package aurora.carevisionapiserver.global.util.validation.validator;
+package aurora.carevisionapiserver.global.auth.util;
 
 import org.springframework.stereotype.Component;
 
 import aurora.carevisionapiserver.global.auth.exception.AuthException;
 import aurora.carevisionapiserver.global.auth.repository.RefreshTokenRepository;
-import aurora.carevisionapiserver.global.auth.util.JWTUtil;
 import aurora.carevisionapiserver.global.error.code.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
 
@@ -14,12 +13,6 @@ public class RefreshTokenValidator {
 
     private final RefreshTokenRepository refreshTokenRepository;
     private final JWTUtil jwtUtil;
-
-    public void checkIfTokenNull(String refreshToken) {
-        if (refreshToken == null) {
-            throw new AuthException(ErrorStatus.REFRESH_TOKEN_NULL);
-        }
-    }
 
     public void validateToken(String refreshToken) {
         jwtUtil.isValidToken(refreshToken);
